Relocate some declarations.

+++ g_phys.c
@@ -909,7 +909,7 @@ void SV_Physics_Step(edict_t* ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_FLY) && (ent->velocity[2] != 0))
 	{
-		speed = fabs(ent->velocity[2]);
+		speed = fabsf(ent->velocity[2]);
 		control = speed < sv_stopspeed ? sv_stopspeed : speed;
 		friction = sv_friction / 3;
 		newspeed = speed - (FRAMETIME * control * friction);
@@ -922,7 +922,7 @@ void SV_Physics_Step(edict_t* ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_SWIM) && (ent->velocity[2] != 0))
 	{
-		speed = fabs(ent->velocity[2]);
+		speed = fabsf(ent->velocity[2]);
 		control = speed < sv_stopspeed ? sv_stopspeed : speed;
 		newspeed = speed - (FRAMETIME * control * sv_waterfriction * ent->waterlevel);
 		if (newspeed < 0)
@@ -1028,7 +1028,7 @@ end:
 }
 
 
-/* from Paril, wall hider detection
+/* from Paril, wall hider detection.
 qboolean SV_RunThink4 (edict_t *ent)
 {
 	float	thinktime4;
